+++ actionpack/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Rails 3.2.6 (unreleased) ##
+
+*   Include mounted_helpers (helpers for accessing mounted engines) in ActionDispatch::IntegrationTest by default. *Piotr Sarnacki*
+
 ## Rails 3.2.4 (May 31, 2012) ##
 
 *   Deprecate old APIs for highlight, excerpt and word_wrap *Jeremy Walker*